Changed way kernel variants are defined, selected to run, and reported.
Now variants are not guarded by compilation defined macros and each
kernel explicitly states which variants it has defined. Only those that
are defined will be run if selected cia input option.

This makes it easy to support kernels that only provide a subset of
variants or variants that apply only to some kernels, not all.