Add base cuda INDEXLIST implementation
This is a somewhat naive implementation where each block
waits for the previous to complete. This may be made better by
reading multiple previous blocks and completing parts of the grid
scan multiple times. It would also help to reduce the number of
communicating blocks by assigning more work to a single block.
Note that blocks are assigned as we come across them with atomics
instead of via blockIdx so blocks only wait on other blocks
that are already running.