Add --gpu_block_size argument, kernels get as part of params
My current thinking is that all kernels will have a default block_size
that will be used if no gpu_block_size is set. If a block_size is set
then any kernels that do not support that block size will not run.
Kernels have a method to indicate if they support a certain block_size
or not, by default kernels only support the default block size.