change Handling of default gpu replications in multi-reductions
Now the default gpu lists are empty.
The default gpu tuning paramaters are now run if atomic replications
are not set and the tuning values are not included in the tuning name.
This allows easier comparison between backends as the
tuning names are now the same in the default case.