change Handling of default items per thread in scans
Now the default gpu lists are empty.
The default gpu tuning paramaters are now run if items per thread
is not set and the tuning values are not included in the tuning name.
This allows easier comparison between backends as the
tuning names are now the same in the default case.