Add counting to most kernels.
Some kernels do not yet support counting and will have values of -1
to indicate that they were not counted.
In some cases the kernels helper file may have not cleanly worked with the
wrappers. Some kernels use library functions like MPI or std::sort that
may not reliably work with the wrappers.