Change way variants and tunings are specified
Now tunings are stored as a KernelBase member function pointer
taking a VariantID.
These can now be specified in a single add*VariantTunings method call
instead of being named in set*TuningDefinitions and run via
run*Variant. This deduplicates helper code and reduces the need
to interact with tuning indices.