Add ChecksumConsistency attribute
This records whether all variant tunings of a kernel always get the
same answer, or if the answer varies between variant tunings but is
consistent for each variant tuning, or if the answer varies from
run to run (atomics).
This attribute is implemented like computational complexity.

I chose values that are consistent with the current tunings. However
some kernels could be implemented with atomics which would move them
to inconsistent. For example kernels with dot products which are
currently implemented as sequential loops even on GPUs.